Added github actions workflows and initial unit tests
Added github actions workflows for a push to a branch,
main/master branch, and when performing a pull request.
We do this in case access to secret variables is
required as they are not available when performing a
pull request. Additionally, there is a separate
workflow for a branch, and main/master branch. This
will allow us to generate a workflow run for
publishing documentation.

We also added the first unit tests comparing the final
resulsts from the MCD C-term scripts. This has been
implemented with a shell script that handles running
the MCD suite and pytest to then compare the data in
the mcd-c-spectrum-? files. We make use of pytest as
this has a very simple way to check the unit tests
and any failures are return an exit code that is
easily understood with github actions.