Some modifications to generate the MChD A terms correctly
The inclusion of the electric quadrupole to calculate the MChD A(A') term has been corrected.

Some debugging prints have been set to verify the correct diagonalisation of the Zeeman Hamiltonian within each set of degenerate states in succession for the calculation of MChD A terms.

Bonus: Correction of the counting of calculated transitions for both MChD A and C terms.